WIP: Logging to File
Started looking into a Logging rework, as in the previous version it was simply an in-game 'text' notification, in a function directly in the `Main` class. Currently a prototype File logging is present, however how this functions will be changed - as I want to allow users to enable the Debugging, instead of them being compiled out - which I would do in a normal application.