Completed Initial Logging
Merged current `configuration-rewrite` as to have up-to-date-enough `Configuration` class, however a rethink of how the `Configuration` or `Logging` class functions may need to happen, as it stands, having to pass the `Configuration` each time a call of the Logging class is called is untenable. `Logger` and/or `Configuration` may need to be a changed to a Singleton. The config should be accessible at all times to the program, however I was keeping it as a regular object to control what has access to this - however this is really just a mod at the end of the day and this isn't really the end of the world.